+++ SIL.ReleaseTasks.Tests/StampChangelogFileWithVersionTests.cs
@@ -11,7 +11,7 @@ namespace SIL.ReleaseTasks.Tests
 	public class StampChangelogFileWithVersionTests
 	{
 		[Test]
-		public void StampMarkdownWorks()
+		public void StampMarkdownWorksWithDefault()
 		{
 			var testMarkdown = new StampChangelogFileWithVersion();
 			using(var tempFiles = new TwoTempFilesForTest(Path.Combine(Path.GetTempPath(), "Test.md"), null))
@@ -20,11 +20,28 @@ namespace SIL.ReleaseTasks.Tests
 					new[] {"## DEV_VERSION_NUMBER: DEV_RELEASE_DATE", "*with some random content", "*does some things"});
 				testMarkdown.ChangelogFile = tempFiles.FirstFile;
 				testMarkdown.VersionNumber = "2.3.10";
-				var day = $"{DateTime.Now:dd/MMM/yyyy}";
 				Assert.That(testMarkdown.Execute(), Is.True);
 				var newContents = File.ReadAllLines(tempFiles.FirstFile);
 				Assert.That(newContents.Length == 3);
-				Assert.That(newContents[0], Is.EqualTo("## 2.3.10 " + day));
+				Assert.That(newContents[0], Is.EqualTo($"## 2.3.10 {DateTime.Now:yyyy-MM-dd}"));
+			}
+		}
+
+		[Test]
+		public void StampMarkdownCustomFormat()
+		{
+			var testMarkdown = new StampChangelogFileWithVersion();
+			using(var tempFiles = new TwoTempFilesForTest(Path.Combine(Path.GetTempPath(), "Test.md"), null))
+			{
+				File.WriteAllLines(tempFiles.FirstFile,
+					new[] {"## DEV_VERSION_NUMBER: DEV_RELEASE_DATE", "*with some random content", "*does some things"});
+				testMarkdown.ChangelogFile = tempFiles.FirstFile;
+				testMarkdown.VersionNumber = "2.3.10";
+				testMarkdown.DateTimeFormat = "dd/MMM/yyyy";
+				Assert.That(testMarkdown.Execute(), Is.True);
+				var newContents = File.ReadAllLines(tempFiles.FirstFile);
+				Assert.That(newContents.Length == 3);
+				Assert.That(newContents[0], Is.EqualTo($"## 2.3.10 {DateTime.Now:dd/MMM/yyyy}"));
 			}
 		}
 	}